+++ R/site_models_to_xml_state.R
@@ -1,3 +1,5 @@
+#' Deprecated function
+#'
 #' Converts one or more clock models to the \code{state} section of the
 #' XML as text
 #' @inheritParams default_params_doc
@@ -6,15 +8,11 @@
 #' @author Richèl J.C. Bilderbeek
 #' @export
 site_models_to_xml_state <- function(
-  site_models
+  site_models = "deprecated"
 ) {
-  testit::assert(beautier::are_site_models(site_models))
+  stop(
+    "Deprecated since beautier v2.5.",
+    "See v2.4 for the code"
+  )
 
-  text <- NULL
-  for (site_model in site_models) {
-    text <- c(text,
-      beautier::site_model_to_xml_state(site_model)
-    )
-  }
-  text
 }