+++ tests/testthat/test-create_beast2_input_beast.R
@@ -1,12 +1,26 @@
-context("create_beast2_input_beast")
-
-test_that("use, one alignment", {
+test_that("use, v2.4, one alignment", {
 
   fasta_filename <- get_beautier_path("anthus_aco.fas")
   inference_model <- init_inference_model(
     input_filename = fasta_filename,
     inference_model = create_inference_model()
   )
+  expect_silent(
+    create_beast2_input_beast(
+      input_filename = fasta_filename,
+      inference_model = inference_model
+    )
+  )
+})
+
+test_that("use, v2.6, one alignment", {
+
+  fasta_filename <- get_beautier_path("anthus_aco.fas")
+  inference_model <- init_inference_model(
+    input_filename = fasta_filename,
+    inference_model = create_inference_model()
+  )
+  inference_model$beauti_options <- create_beauti_options_v2_6()
   expect_silent(
     create_beast2_input_beast(
       input_filename = fasta_filename,