+++ tests/testthat/test-create_beast2_input_state.R
@@ -20,6 +20,17 @@ test_that("v2.4", {
   expect_equal(created, expected)
 })
 
+test_that("v2.6", {
+
+  inference_model <- init_inference_model(
+    input_filename = get_fasta_filename(),
+    inference_model = create_test_inference_model(
+      beauti_options = create_beauti_options_v2_6()
+    )
+  )
+  expect_silent(create_beast2_input_state(inference_model = inference_model))
+})
+
 test_that("birth_death", {
 
   skip("Obsoleted interface")