+++ tests/testthat/test-create_site_model_parameters_xml.R
@@ -59,7 +59,6 @@ test_that("use, JC69, v2.4, GCC = 2", {
 })
 
 test_that("use, JC69, v2.6", {
-  skip("WIP")
   inference_model <- init_inference_model(
     input_filename = get_fasta_filename(),
     inference_model = create_test_inference_model(