+++ tests/testthat/test-site_models_to_xml_state.R
@@ -1,0 +1,6 @@
+test_that("use", {
+  expect_equal(
+    site_models_to_xml_state(),
+    "Deprecated"
+  )
+})